fix(context): manager - move directive into is component fix(context): change /profils-users for /profils/users
(cherry picked from commit a71d6ce2db1353d643469389500f52bf1ca9180a)
