+++ packages/context/src/lib/context-manager/context-manager.directive.ts
@@ -2,10 +2,7 @@ import { ContextEditComponent } from './context-edit';
 import { ContextFormComponent } from './context-form';
 import { ContextItemComponent } from './context-item';
 import { ContextListComponent } from './context-list';
-import {
-  ContextPermissionsBindingDirective,
-  ContextPermissionsComponent
-} from './context-permissions';
+import { ContextPermissionsComponent } from './context-permissions';
 import { LayerContextDirective, MapContextDirective } from './shared';
 
 export const CONTEXT_MANAGER_DIRECTIVES = [
@@ -14,7 +11,6 @@ export const CONTEXT_MANAGER_DIRECTIVES = [
   ContextFormComponent,
   ContextEditComponent,
   ContextPermissionsComponent,
-  ContextPermissionsBindingDirective,
   MapContextDirective,
   LayerContextDirective
 ] as const;
