+++ packages/context/src/lib/context-manager/context-manager.module.ts
@@ -4,7 +4,6 @@ import { ContextEditComponent } from './context-edit/context-edit.component';
 import { ContextFormComponent } from './context-form/context-form.component';
 import { ContextItemComponent } from './context-item/context-item.component';
 import { ContextListComponent } from './context-list/context-list.component';
-import { ContextPermissionsBindingDirective } from './context-permissions/context-permissions-binding.directive';
 import { ContextPermissionsComponent } from './context-permissions/context-permissions.component';
 import { LayerContextDirective } from './shared/layer-context.directive';
 import { MapContextDirective } from './shared/map-context.directive';
@@ -19,7 +18,6 @@ import { MapContextDirective } from './shared/map-context.directive';
     ContextFormComponent,
     ContextEditComponent,
     ContextPermissionsComponent,
-    ContextPermissionsBindingDirective,
     LayerContextDirective,
     MapContextDirective
   ],
@@ -29,7 +27,6 @@ import { MapContextDirective } from './shared/map-context.directive';
     ContextFormComponent,
     ContextEditComponent,
     ContextPermissionsComponent,
-    ContextPermissionsBindingDirective,
     LayerContextDirective,
     MapContextDirective
   ]
