--- packages/context/src/lib/context-manager/context-permissions/context-permissions-binding.directive.spec.ts
@@ -1,14 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-describe('ContextPermissionsBindingDirective', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      imports: [],
-      providers: []
-    });
-  });
-
-  it('should create an instance', () => {
-    expect(true).toBeTruthy();
-  });
-});
