+++ packages/context/src/lib/context-manager/context-permissions/index.ts
@@ -1,2 +1,1 @@
 export * from './context-permissions.component';
-export * from './context-permissions-binding.directive';
