+++ packages/integration/src/lib/context/context-permission-manager-tool/context-permission-manager-tool.component.html
@@ -1,1 +1,1 @@
-<igo-context-permissions igoContextPermissionsBinding />
+<igo-context-permissions />
