+++ packages/integration/src/lib/context/context-permission-manager-tool/context-permission-manager-tool.component.ts
@@ -1,10 +1,7 @@
 import { Component } from '@angular/core';
 
 import { ToolComponent } from '@igo2/common/tool';
-import {
-  ContextPermissionsBindingDirective,
-  ContextPermissionsComponent
-} from '@igo2/context';
+import { ContextPermissionsComponent } from '@igo2/context';
 
 @ToolComponent({
   name: 'contextPermissionManager',
@@ -15,6 +12,6 @@ import {
 @Component({
   selector: 'igo-context-permission-manager-tool',
   templateUrl: './context-permission-manager-tool.component.html',
-  imports: [ContextPermissionsComponent, ContextPermissionsBindingDirective]
+  imports: [ContextPermissionsComponent]
 })
 export class ContextPermissionManagerToolComponent {}
